+++ packages/adminui-vue/locale/lang-ky.js
@@ -57,6 +57,12 @@ UB.i18nExtend({
         oct: 'Окт',
         nov: 'Ноя',
         dec: 'Дек'
+      },
+      format: {
+        date: 'MM/dd/yyyy',
+        daterange: 'MM/dd/yyyy',
+        datetime: 'MM/dd/yyyy HH:mm',
+        firstDayOfWeek: 7
       }
     },
     select: {